cleanup duplicates pipes + move all pipes to the same place + created a pipes.module.ts
