--- alv-portal-ui/src/app/shared/layout/pipes/phone-number.pipe.ts
@@ -1,16 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-import { CountryCode, format, isValidNumber, parse, } from 'libphonenumber-js';
-
-@Pipe({
-  name: 'jr2PhoneNumber'
-})
-export class PhoneNumberPipe implements PipeTransform {
-
-  transform(value: string, country: CountryCode = 'CH'): string {
-    if (value && isValidNumber(value, country)) {
-      const parsedNumber = parse(value, country);
-      return format(parsedNumber, 'International');
-    }
-    return value;
-  }
-}
