+++ alv-portal-ui/src/app/shared/pipes/capitalize.pipe.ts
