+++ alv-portal-ui/src/app/shared/pipes/enum-keys.pipe.ts
