+++ alv-portal-ui/src/app/shared/pipes/locale-aware-number.pipe.ts
