+++ alv-portal-ui/src/app/shared/pipes/pipes-util.ts
