+++ alv-portal-ui/src/app/shared/pipes/pipes.module.ts
@@ -1,0 +1,45 @@
+import { NgModule } from '@angular/core';
+import { SafeHtmlPipe } from './safe-html.pipe';
+import { LocaleAwareDatePipe } from './locale-aware-date.pipe';
+import { PhoneNumberPipe } from './phone-number.pipe';
+import { CapitalizePipe } from './capitalize.pipe';
+import { KeysPipe } from './enum-keys.pipe';
+import { LocaleAwareDecimalPipe } from './locale-aware-number.pipe';
+import { MarkdownModule } from 'ngx-markdown';
+import { ShortenPipe } from './shorten.pipe';
+import { MarkdownEscapePipe } from './markdown-escape.pipe';
+import { WorkingTimeRangePipe } from './working-time-range.pipe';
+
+@NgModule({
+  declarations: [
+    SafeHtmlPipe,
+    LocaleAwareDatePipe,
+    WorkingTimeRangePipe,
+    PhoneNumberPipe,
+    CapitalizePipe,
+    KeysPipe,
+    LocaleAwareDecimalPipe,
+    MarkdownEscapePipe,
+    ShortenPipe
+  ],
+  imports: [
+    MarkdownModule.forRoot()
+  ],
+  entryComponents: [],
+  exports: [
+    SafeHtmlPipe,
+    LocaleAwareDatePipe,
+    WorkingTimeRangePipe,
+    PhoneNumberPipe,
+    CapitalizePipe,
+    KeysPipe,
+    LocaleAwareDecimalPipe,
+    MarkdownEscapePipe,
+    ShortenPipe
+  ]
+})
+export class PipesModule {
+
+}
+
+
