+++ alv-portal-ui/src/app/shared/pipes/shorten.pipe.ts
